Revert changes to InvulnerabilityHandler
